Fix setup-dev

+++ bin/setup-dev.sh
@@ -1,17 +1,29 @@
 #!/usr/bin/env bash
 
-cd ~/
 
-wget http://dlib.net/files/dlib-19.6.tar.bz2
-tar xvf dlib-19.6.tar.bz2
-rm -f dlib-19.6.tar.bz2
+# CentOS
+if [ -f /etc/redhat-release ]; then
+    echo "Installation for Centos not found!"
+fi
 
-cd dlib-19.6/
-mkdir build
-cd build
 
-cmake ..
-cmake --build . --config Release
-sudo make install
-sudo ldconfig
-cd ..
+# Ubuntu
+if [ -f /etc/lsb-release ]; then
+
+    cd ~/
+
+    wget http://dlib.net/files/dlib-19.6.tar.bz2
+    tar xvf dlib-19.6.tar.bz2
+    rm -f dlib-19.6.tar.bz2
+
+    cd dlib-19.6/
+    mkdir build
+    cd build
+
+    cmake ..
+    cmake --build . --config Release
+    sudo make install
+    sudo ldconfig
+    cd ..
+
+fi
